requirements.txt
PyYAML~=6.0.1
python-dotenv~=1.0.1
pyprojroot~=0.3.0
openai~=1.37.1
gradio~=4.40.0
langchain-community~=0.2.10
langchain-core~=0.2.26
langchain~=0.2.11
neo4j~=5.23.0
langchain-openai~=0.1.20

# pip install pandas
# pip install seaborn
# pip install langchain-openai